Fix IdentityExtensions
